+++ tests/UnixFileMappingReaderTest.php
@@ -24,25 +24,41 @@ class UnixFileMappingReaderTest extends TestCase
             sha1(__METHOD__),
             null,
             [
-                'files' => '{foo,bar}.php',
-                'source' => [
-                    'foo.php' => 'Foo'
-                ],
+                'files' =>
+                    "{foo,bar}.php\n{templates/dot,.}gitignore:merge:force\n",
+                'source' => [],
                 'destination' => []
             ]
         );
 
-        $mappings = new UnixFileMappingReader(
-            $fileSystem->getChild('source')->url(),
-            $fileSystem->getChild('destination')->url(),
-            $fileSystem->getChild('files')->url(),
+        $sourceDirectory = $fileSystem->getChild('source')->url();
+        $destinationDirectory = $fileSystem->getChild('destination')->url();
+        $mappingsReader = new UnixFileMappingReader(
+            $sourceDirectory,
+            $destinationDirectory,
             $fileSystem->getChild('files')->url()
         );
 
-        foreach ($mappings as $offset => $mapping) {
-            $this->assertInstanceOf(FileMappingInterface::class, $mapping);
-            $this->assertIsInt($offset);
-            $this->assertFileExists($mapping->getSource());
-        }
+        /** @var FileMappingInterface[] $mappings */
+        $mappings = iterator_to_array($mappingsReader);
+
+        $this->assertIsList($mappings);
+        $this->assertCount(2, $mappings);
+
+        // Verify mapping '{foo,bar}.php'
+        $this->assertInstanceOf(FileMappingInterface::class, $mappings[0]);
+        $this->assertSame('foo.php', $mappings[0]->getRelativeSource());
+        $this->assertSame($sourceDirectory . DIRECTORY_SEPARATOR . 'foo.php', $mappings[0]->getSource());
+        $this->assertSame('bar.php', $mappings[0]->getRelativeDestination());
+        $this->assertSame($destinationDirectory . DIRECTORY_SEPARATOR . 'bar.php', $mappings[0]->getDestination());
+        $this->assertSame([], $mappings[0]->getOptions());
+
+        // Verify mapping '{templates/dot,.}gitignore:merge:force'
+        $this->assertInstanceOf(FileMappingInterface::class, $mappings[1]);
+        $this->assertSame('templates/dotgitignore', $mappings[1]->getRelativeSource());
+        $this->assertSame($sourceDirectory . DIRECTORY_SEPARATOR . 'templates/dotgitignore', $mappings[1]->getSource());
+        $this->assertSame('.gitignore', $mappings[1]->getRelativeDestination());
+        $this->assertSame($destinationDirectory . DIRECTORY_SEPARATOR . '.gitignore', $mappings[1]->getDestination());
+        $this->assertSame(['merge', 'force'], $mappings[1]->getOptions());
     }
 }